+++ src/ceno_binding/error.rs
