+++ src/lib.rs
@@ -1,7 +1,8 @@
+#[cfg(feature = "ceno")]
+pub mod ceno_binding; // Connect whir with ceno
 pub mod cmdline_utils;
 pub mod crypto; // Crypto utils
 pub mod domain; // Domain that we are evaluating over
-pub mod errors;
 pub mod fs_utils;
 pub mod ntt;
 pub mod parameters;
