+++ src/whir/mod.rs
@@ -1,16 +1,13 @@
 use ark_crypto_primitives::merkle_tree::{Config, MultiPath};
-use ark_ff::FftField;
 use ark_serialize::{CanonicalDeserialize, CanonicalSerialize};
 use std::fmt::Debug;
 
-use crate::errors::Error;
 use crate::poly_utils::MultilinearPoint;
 
 pub mod committer;
 mod fs_utils;
 pub mod iopattern;
 pub mod parameters;
-pub mod pcs;
 pub mod prover;
 pub mod verifier;
 
@@ -36,64 +33,6 @@ where
     F: Sized + Clone + CanonicalSerialize + CanonicalDeserialize,
 {
     transcript.len() + whir_proof.serialized_size(ark_serialize::Compress::Yes)
-}
-
-pub trait PolynomialCommitmentScheme<E: FftField>: Clone {
-    type Param: Clone;
-    type CommitmentWithData;
-    type Proof: Clone + CanonicalSerialize + CanonicalDeserialize;
-    type Poly: Clone;
-    type Transcript;
-
-    fn setup(poly_size: usize) -> Self::Param;
-
-    fn commit_and_write(
-        pp: &Self::Param,
-        poly: &Self::Poly,
-        transcript: &mut Self::Transcript,
-    ) -> Result<Self::CommitmentWithData, Error>;
-
-    fn batch_commit(
-        pp: &Self::Param,
-        polys: &[Self::Poly],
-    ) -> Result<Self::CommitmentWithData, Error>;
-
-    fn open(
-        pp: &Self::Param,
-        comm: Self::CommitmentWithData,
-        point: &[E],
-        eval: &E,
-        transcript: &mut Self::Transcript,
-    ) -> Result<Self::Proof, Error>;
-
-    /// This is a simple version of batch open:
-    /// 1. Open at one point
-    /// 2. All the polynomials share the same commitment.
-    /// 3. The point is already a random point generated by a sum-check.
-    fn batch_open(
-        pp: &Self::Param,
-        polys: &[Self::Poly],
-        comm: Self::CommitmentWithData,
-        point: &[E],
-        evals: &[E],
-        transcript: &mut Self::Transcript,
-    ) -> Result<Self::Proof, Error>;
-
-    fn verify(
-        vp: &Self::Param,
-        point: &[E],
-        eval: &E,
-        proof: &Self::Proof,
-        transcript: &Self::Transcript,
-    ) -> Result<(), Error>;
-
-    fn batch_verify(
-        vp: &Self::Param,
-        point: &[E],
-        evals: &[E],
-        proof: &Self::Proof,
-        transcript: &mut Self::Transcript,
-    ) -> Result<(), Error>;
 }
 
 #[cfg(test)]
